+++ src/Terrabuild/SourceControls/Local.fs
@@ -4,9 +4,11 @@ open Environment
 type Local() =
     let commitLog = currentDir() |> Git.getCommitLog
     let commit = commitLog.Head
+    let repository = currentDir() |> Git.tryGetOriginRemote |> Option.defaultValue ""
 
     interface Contracts.ISourceControl with
         override _.BranchOrTag = currentDir() |> Git.getBranchOrTag
+        override _.Repository = repository
         
         override _.HeadCommit =
             { Sha = commit.Sha