+++ src/Terrabuild/Web/GraphServer.fs
@@ -170,6 +170,7 @@ let private buildConfig
               ConfigOptions.Options.HeadCommit = sourceControl.HeadCommit
               ConfigOptions.Options.CommitLog = sourceControl.CommitLog
               ConfigOptions.Options.BranchOrTag = sourceControl.BranchOrTag
+              ConfigOptions.Options.Repository = sourceControl.Repository
               ConfigOptions.Options.Run = sourceControl.Run }
         Configuration.read options
     finally